+++ tools/highlighter/factory.cpp
@@ -78,6 +78,9 @@ FormatterFactory::create(std::ostream &stream) const {
     if (this->lineno) {
       setFlag(formatOp, HTMLFormatOp::LINENO);
     }
+    if (this->htmlTable) {
+      setFlag(formatOp, HTMLFormatOp::TABLE | HTMLFormatOp::LINENO);
+    }
     unsigned int lineNumOffset = 1;
     if (this->lineno) {
       auto ret = convertToNum<unsigned int>(this->lineno);